pyproject.toml
[project]
name = "auto-quant-tool"
version = "0.1.0"
description = "Automated quantization benchmarking suite for GGUF, GPTQ, and TFLite models"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12",
    "pydantic>=2.0",
    "huggingface_hub>=0.23",
    "rich>=13.0",
    "pyyaml>=6.0.3",
    "numpy>=2.2.6",
    "llama-cpp-python>=0.3.20",
    "onnx>=1.21.0",
    "onnxscript>=0.6.2",
    "plotly>=6.6.0",
    "kaleido>=1.2.0",
    "gradio>=6.11.0",
    "pandas>=2.3.3",
]

[project.optional-dependencies]
gguf      = ["llama-cpp-python"]
gptq      = ["auto-gptq"]
tflite    = ["ai-edge-torch"]
ui        = ["gradio>=4.0"]
benchmark = [
    "scipy>=1.13",
    "plotly>=5.0",
    "pandas>=2.0",
    "datasets>=2.0",
]

[project.scripts]
auto-quant = "auto_quant.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
exclude = [
    "third_party/",
    "outputs/",
    "notebooks/",
    "*.yaml",
    ".venv/",
]
"*" = []